confining potentials
